+++ src/main/java/com/woowacourse/zzinbros/post/exception/PostException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzinbros.post.domain;
+package com.woowacourse.zzinbros.post.exception;
 
 public class PostException extends RuntimeException {
     public PostException() {
